Add authorization to resource capacity

+++ app/Http/Controllers/Resource/ResourceCapacityController.php
@@ -3,34 +3,33 @@
 namespace App\Http\Controllers\Resource;
 
 use App\ResourceCapacity;
-use Illuminate\Http\Request;
+use App\Http\Requests\Resource\ResourceCapacityRequest;
 use App\Http\Controllers\Controller;
 use App\Resource;
-use TiMacDonald\Validation\Rule;
 use Illuminate\Support\Facades\Redirect;
 use Carbon\Carbon;
-use App\Rules\NoIntervalOverlap;
-use App\Libraries\DateHelper;
-use App\Model;
 
 class ResourceCapacityController extends Controller
 {
     public function index(Resource $resource)
     {
+        $this->authorize('viewAny', ResourceCapacity::class);
         $capacities = $resource->capacities()->ordered()->get();
         return view('resources.resources.capacities.index', compact('resource', 'capacities'));
     }
 
     public function create(Resource $resource)
     {
+        $this->authorize('create', ResourceCapacity::class);
         $resourceCapacity = new ResourceCapacity(['start' => Carbon::now()->setDay(1), 'end' => Carbon::now()->addYear()->setDay(1)]);
         $formAction = route('resources.capacities.store', ['resource' => $resource->pid]);
         return view('resources.resources.capacities.edit', compact('resource', 'resourceCapacity', 'formAction'));
     }
 
-    public function store(Request $request, Resource $resource)
+    public function store(ResourceCapacityRequest $request, Resource $resource)
     {
-        $capacity = new ResourceCapacity($this->validateValues($request));
+        $this->authorize('create', ResourceCapacity::class);
+        $capacity = new ResourceCapacity($request->validated());
         $capacity->resource_id = $resource->id;
         $capacity->save();
         return Redirect::route('resources.capacities.index', ['resource' => $resource->pid]);
@@ -38,6 +37,7 @@ class ResourceCapacityController extends Controller
 
     public function edit(Resource $resource, ResourceCapacity $resourceCapacity)
     {
+        $this->authorize('view', $resourceCapacity);
         if ($resourceCapacity->resource_id != $resource->id) {
             abort(404);
         }
@@ -45,34 +45,23 @@ class ResourceCapacityController extends Controller
         return view('resources.resources.capacities.edit', compact('resource', 'resourceCapacity', 'formAction'));
     }
 
-    public function update(Request $request, Resource $resource, ResourceCapacity $resourceCapacity)
+    public function update(ResourceCapacityRequest $request, Resource $resource, ResourceCapacity $resourceCapacity)
     {
+        $this->authorize('update', $resourceCapacity);
         if ($resourceCapacity->resource_id != $resource->id) {
             abort(404);
         }
-        $resourceCapacity->update($this->validateValues($request));
+        $resourceCapacity->update($request->validated());
         return Redirect::route('resources.capacities.index', ['resource' => $resource->pid]);
     }
 
     public function destroy(Resource $resource, ResourceCapacity $resourceCapacity)
     {
+        $this->authorize('delete', $resourceCapacity);
         if ($resourceCapacity->resource_id != $resource->id) {
             abort(404);
         }
         $resourceCapacity->delete();
         return Redirect::route('resources.capacities.index', ['resource' => $resource->pid]);
-    }
-
-    private function validateValues(Request $request)
-    {
-        $values = $request->validate([
-            'start' => Rule::required()->after(Model::DD_DATE_MIN)->before(Model::DD_DATE_MAX)->get(),
-            'end' => Rule::required()->after(Model::DD_DATE_MIN)->before(Model::DD_DATE_MAX)->after('start')->add(new NoIntervalOverlap($request))->get(),
-            'type' => Rule::required()->in(array_keys(ResourceCapacity::TYPES))->get(),
-            'quantity' => Rule::required()->integer(0, ResourceCapacity::DD_QUANTITY_MAX)->get()
-        ]);
-        $values['start'] = DateHelper::setDateToMonth($values['start']);
-        $values['end'] = DateHelper::setDateToMonth($values['end']);
-        return $values;
     }
 }
